api/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.11
h11==0.14.0
idna==3.10
psutil==6.1.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
redis==5.2.1
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
PyJWT~=2.10.1
passlib~=1.7.4
sqlalchemy~=2.0.39
psycopg2~=2.9.10
python-multipart~=0.0.20
pandas~=2.2.3